WIP Visitor seems now to work correctly. Split out types used by ast into slang-ast-support-types.h
