+++ source/slang/slang-ast-decl.h
@@ -6,6 +6,9 @@
 
 namespace Slang {
 
+#define SLANG_ABSTRACT_CLASS(x) SLANG_ABSTRACT_CLASS_REFLECT(x)
+#define SLANG_CLASS(x) SLANG_CLASS_REFLECT_WITH_ACCEPT(x)
+
 // Syntax class definitions for declarations.
 
 // A group of declarations that should be treated as a unit
@@ -455,5 +458,7 @@ class AttributeDecl : public ContainerDecl
     SyntaxClass<RefObject> syntaxClass;
 };
 
+#undef SLANG_ABSTRACT_CLASS
+#undef SLANG_CLASS
 
 } // namespace Slang
