+++ source/slang/slang-ast-modifier.h
@@ -6,28 +6,31 @@
 
 namespace Slang { 
 
+#define SLANG_ABSTRACT_CLASS(x) SLANG_ABSTRACT_CLASS_REFLECT(x)
+#define SLANG_CLASS(x) SLANG_CLASS_REFLECT_WITH_ACCEPT(x)
+
 // Syntax class definitions for modifiers.
 
 // Simple modifiers have no state beyond their identity
 
-class InModifier : public Modifier { SLANG_CLASS(InModifier) };
-class OutModifier : public Modifier { SLANG_CLASS(OutModifier) };
-class ConstModifier : public Modifier { SLANG_CLASS(ConstModifier) };
-class InstanceModifier : public Modifier { SLANG_CLASS(InstanceModifier) };
-class BuiltinModifier : public Modifier { SLANG_CLASS(BuiltinModifier) };
-class InlineModifier : public Modifier { SLANG_CLASS(InlineModifier) };
-class PublicModifier : public Modifier { SLANG_CLASS(PublicModifier) };
-class RequireModifier : public Modifier { SLANG_CLASS(RequireModifier) };
-class ParamModifier : public Modifier { SLANG_CLASS(ParamModifier) };
-class ExternModifier : public Modifier { SLANG_CLASS(ExternModifier) };
-class InputModifier : public Modifier { SLANG_CLASS(InputModifier) };
-class TransparentModifier : public Modifier { SLANG_CLASS(TransparentModifier) };
-class FromStdLibModifier : public Modifier { SLANG_CLASS(FromStdLibModifier) };
-class PrefixModifier : public Modifier { SLANG_CLASS(PrefixModifier) };
-class PostfixModifier : public Modifier { SLANG_CLASS(PostfixModifier) };
-class ExportedModifier : public Modifier { SLANG_CLASS(ExportedModifier) };
-class ConstExprModifier : public Modifier { SLANG_CLASS(ConstExprModifier) };
-class GloballyCoherentModifier : public Modifier { SLANG_CLASS(GloballyCoherentModifier) };
+class InModifier : public Modifier { SLANG_CLASS(InModifier)};
+class OutModifier : public Modifier { SLANG_CLASS(OutModifier)};
+class ConstModifier : public Modifier { SLANG_CLASS(ConstModifier)};
+class InstanceModifier : public Modifier { SLANG_CLASS(InstanceModifier)};
+class BuiltinModifier : public Modifier { SLANG_CLASS(BuiltinModifier)};
+class InlineModifier : public Modifier { SLANG_CLASS(InlineModifier)};
+class PublicModifier : public Modifier { SLANG_CLASS(PublicModifier)};
+class RequireModifier : public Modifier { SLANG_CLASS(RequireModifier)};
+class ParamModifier : public Modifier { SLANG_CLASS(ParamModifier)};
+class ExternModifier : public Modifier { SLANG_CLASS(ExternModifier)};
+class InputModifier : public Modifier { SLANG_CLASS(InputModifier)};
+class TransparentModifier : public Modifier { SLANG_CLASS(TransparentModifier)};
+class FromStdLibModifier : public Modifier { SLANG_CLASS(FromStdLibModifier)};
+class PrefixModifier : public Modifier { SLANG_CLASS(PrefixModifier)};
+class PostfixModifier : public Modifier { SLANG_CLASS(PostfixModifier)};
+class ExportedModifier : public Modifier { SLANG_CLASS(ExportedModifier)};
+class ConstExprModifier : public Modifier { SLANG_CLASS(ConstExprModifier)};
+class GloballyCoherentModifier : public Modifier { SLANG_CLASS(GloballyCoherentModifier)};
 
 // A modifier that marks something as an operation that
 // has a one-to-one translation to the IR, and thus
@@ -44,7 +47,7 @@ class IntrinsicOpModifier : public Modifier
     // The IR opcode for the intrinsic operation.
     //
     uint32_t op = 0;
- };
+};
 
 // A modifier that marks something as an intrinsic function,
 // for some subset of targets.
@@ -58,7 +61,7 @@ class TargetIntrinsicModifier : public Modifier
 
     // A custom definition for the operation
     Token definitionToken;
- };
+};
 
 // A modifier that marks a declaration as representing a
 // specialization that should be preferred on a particular
@@ -70,7 +73,7 @@ class SpecializedForTargetModifier : public Modifier
     // Token that names the target that the operation
     // has been specialized for.
     Token targetToken;
- };
+};
 
 // A modifier to tag something as an intrinsic that requires
 // a certain GLSL extension to be enabled when used
@@ -79,7 +82,7 @@ class RequiredGLSLExtensionModifier : public Modifier
     SLANG_CLASS(RequiredGLSLExtensionModifier)
  
     Token extensionNameToken;
- };
+};
 
 // A modifier to tag something as an intrinsic that requires
 // a certain GLSL version to be enabled when used
@@ -88,7 +91,7 @@ class RequiredGLSLVersionModifier : public Modifier
     SLANG_CLASS(RequiredGLSLVersionModifier)
  
     Token versionNumberToken;
- };
+};
 
 
 // A modifier to tag something as an intrinsic that requires
@@ -98,7 +101,7 @@ class RequiredSPIRVVersionModifier : public Modifier
     SLANG_CLASS(RequiredSPIRVVersionModifier)
  
     SemanticVersion version;
- };
+};
 
 // A modifier to tag something as an intrinsic that requires
 // a certain CUDA SM version to be enabled when used. Specified as "major.minor"
@@ -107,19 +110,19 @@ class RequiredCUDASMVersionModifier : public Modifier
     SLANG_CLASS(RequiredCUDASMVersionModifier)
  
     SemanticVersion version;
- };
+};
 
 class InOutModifier : public OutModifier 
 {
     SLANG_CLASS(InOutModifier)
- };
+};
 
 
 // `__ref` modifier for by-reference parameter passing
 class RefModifier : public Modifier 
 {
     SLANG_CLASS(RefModifier)
- };
+};
 
 
 // This is a special sentinel modifier that gets added
@@ -139,7 +142,7 @@ class RefModifier : public Modifier
 class SharedModifiers : public Modifier 
 {
     SLANG_CLASS(SharedModifiers)
- };
+};
 
 
 // A GLSL `layout` modifier
@@ -162,23 +165,23 @@ class GLSLLayoutModifier : public Modifier
 
     // TODO: may want to accept a full expression here
     Token valToken;
- };
+};
 
 // AST nodes to represent the begin/end of a `layout` modifier group
 class GLSLLayoutModifierGroupMarker : public Modifier 
 {
     SLANG_ABSTRACT_CLASS(GLSLLayoutModifierGroupMarker)
- };
+};
 
 class GLSLLayoutModifierGroupBegin : public GLSLLayoutModifierGroupMarker 
 {
     SLANG_CLASS(GLSLLayoutModifierGroupBegin)
- };
+};
 
 class GLSLLayoutModifierGroupEnd : public GLSLLayoutModifierGroupMarker 
 {
     SLANG_CLASS(GLSLLayoutModifierGroupEnd)
- };
+};
 
 
 // We divide GLSL `layout` modifiers into those we have parsed
@@ -187,53 +190,53 @@ class GLSLLayoutModifierGroupEnd : public GLSLLayoutModifierGroupMarker
 class GLSLParsedLayoutModifier : public GLSLLayoutModifier 
 {
     SLANG_ABSTRACT_CLASS(GLSLParsedLayoutModifier)
- };
+};
 
 class GLSLUnparsedLayoutModifier : public GLSLLayoutModifier 
 {
     SLANG_CLASS(GLSLUnparsedLayoutModifier)
- };
+};
 
 
 // Specific cases for known GLSL `layout` modifiers that we need to work with
 class GLSLConstantIDLayoutModifier : public GLSLParsedLayoutModifier 
 {
     SLANG_CLASS(GLSLConstantIDLayoutModifier)
- };
+};
 
 class GLSLLocationLayoutModifier : public GLSLParsedLayoutModifier 
 {
     SLANG_CLASS(GLSLLocationLayoutModifier)
- };
+};
 
 
 // A catch-all for single-keyword modifiers
 class SimpleModifier : public Modifier 
 {
     SLANG_CLASS(SimpleModifier)
- };
+};
 
 
 // Some GLSL-specific modifiers
 class GLSLBufferModifier : public SimpleModifier 
 {
     SLANG_CLASS(GLSLBufferModifier)
- };
+};
 
 class GLSLWriteOnlyModifier : public SimpleModifier 
 {
     SLANG_CLASS(GLSLWriteOnlyModifier)
- };
+};
 
 class GLSLReadOnlyModifier : public SimpleModifier 
 {
     SLANG_CLASS(GLSLReadOnlyModifier)
- };
+};
 
 class GLSLPatchModifier : public SimpleModifier 
 {
     SLANG_CLASS(GLSLPatchModifier)
- };
+};
 
 
 // Indicates that this is a variable declaration that corresponds to
@@ -241,7 +244,7 @@ class GLSLPatchModifier : public SimpleModifier
 class ImplicitParameterGroupVariableModifier : public Modifier 
 {
     SLANG_CLASS(ImplicitParameterGroupVariableModifier)
- };
+};
 
 
 // Indicates that this is a type that corresponds to the element
@@ -249,7 +252,7 @@ class ImplicitParameterGroupVariableModifier : public Modifier
 class ImplicitParameterGroupElementTypeModifier : public Modifier 
 {
     SLANG_CLASS(ImplicitParameterGroupElementTypeModifier)
- };
+};
 
 
 // An HLSL semantic
@@ -258,7 +261,7 @@ class HLSLSemantic : public Modifier
     SLANG_ABSTRACT_CLASS(HLSLSemantic)
  
     Token name;
- };
+};
 
 // An HLSL semantic that affects layout
 class HLSLLayoutSemantic : public HLSLSemantic 
@@ -267,7 +270,7 @@ class HLSLLayoutSemantic : public HLSLSemantic
  
     Token registerName;
     Token componentMask;
- };
+};
 
 // An HLSL `register` semantic
 class HLSLRegisterSemantic : public HLSLLayoutSemantic 
@@ -275,20 +278,20 @@ class HLSLRegisterSemantic : public HLSLLayoutSemantic
     SLANG_CLASS(HLSLRegisterSemantic)
  
     Token spaceName;
- };
+};
 
 // TODO(tfoley): `packoffset`
 class HLSLPackOffsetSemantic : public HLSLLayoutSemantic 
 {
     SLANG_CLASS(HLSLPackOffsetSemantic)
- };
+};
 
 
 // An HLSL semantic that just associated a declaration with a semantic name
 class HLSLSimpleSemantic : public HLSLSemantic 
 {
     SLANG_CLASS(HLSLSimpleSemantic)
- };
+};
 
 
 // GLSL
@@ -298,7 +301,7 @@ class HLSLSimpleSemantic : public HLSLSemantic
 class GLSLPreprocessorDirective : public Modifier 
 {
     SLANG_CLASS(GLSLPreprocessorDirective)
- };
+};
 
 
 // A GLSL `#version` directive
@@ -312,7 +315,7 @@ class GLSLVersionDirective : public GLSLPreprocessorDirective
 
     // Optional token giving the sub-profile to be used
     Token glslProfileToken;
- };
+};
 
 // A GLSL `#extension` directive
 class GLSLExtensionDirective : public GLSLPreprocessorDirective 
@@ -325,14 +328,14 @@ class GLSLExtensionDirective : public GLSLPreprocessorDirective
 
     // Optional token giving the sub-profile to be used
     Token dispositionToken;
- };
+};
 
 class ParameterGroupReflectionName : public Modifier 
 {
     SLANG_CLASS(ParameterGroupReflectionName)
  
     NameLoc nameAndLoc;
- };
+};
 
 // A modifier that indicates a built-in base type (e.g., `float`)
 class BuiltinTypeModifier : public Modifier 
@@ -340,7 +343,7 @@ class BuiltinTypeModifier : public Modifier
     SLANG_CLASS(BuiltinTypeModifier)
  
     BaseType tag;
- };
+};
 
 // A modifier that indicates a built-in type that isn't a base type (e.g., `vector`)
 //
@@ -351,7 +354,7 @@ class MagicTypeModifier : public Modifier
  
     String name;
     uint32_t tag;
- };
+};
 
 // A modifier applied to declarations of builtin types to indicate how they
 // should be lowered to the IR.
@@ -368,37 +371,37 @@ class IntrinsicTypeModifier : public Modifier
     // Additional literal opreands to provide when creating instances.
     // (e.g., for a texture type this passes in shape/mutability info)
     List<uint32_t> irOperands;
- };
+};
 
 // Modifiers that affect the storage layout for matrices
 class MatrixLayoutModifier : public Modifier 
 {
     SLANG_CLASS(MatrixLayoutModifier)
- };
+};
 
 
 // Modifiers that specify row- and column-major layout, respectively
 class RowMajorLayoutModifier : public MatrixLayoutModifier 
 {
     SLANG_CLASS(RowMajorLayoutModifier)
- };
+};
 
 class ColumnMajorLayoutModifier : public MatrixLayoutModifier 
 {
     SLANG_CLASS(ColumnMajorLayoutModifier)
- };
+};
 
 
 // The HLSL flavor of those modifiers
 class HLSLRowMajorLayoutModifier : public RowMajorLayoutModifier 
 {
     SLANG_CLASS(HLSLRowMajorLayoutModifier)
- };
+};
 
 class HLSLColumnMajorLayoutModifier : public ColumnMajorLayoutModifier 
 {
     SLANG_CLASS(HLSLColumnMajorLayoutModifier)
- };
+};
 
 
 // The GLSL flavor of those modifiers
@@ -411,12 +414,12 @@ class HLSLColumnMajorLayoutModifier : public ColumnMajorLayoutModifier
 class GLSLRowMajorLayoutModifier : public ColumnMajorLayoutModifier 
 {
     SLANG_CLASS(GLSLRowMajorLayoutModifier)
- };
+};
 
 class GLSLColumnMajorLayoutModifier : public RowMajorLayoutModifier 
 {
     SLANG_CLASS(GLSLColumnMajorLayoutModifier)
- };
+};
 
 
 // More HLSL Keyword
@@ -425,48 +428,48 @@ class InterpolationModeModifier : public Modifier
 {
     SLANG_ABSTRACT_CLASS(InterpolationModeModifier)
  
- };
+};
 
 // HLSL `nointerpolation` modifier
 class HLSLNoInterpolationModifier : public InterpolationModeModifier 
 {
     SLANG_CLASS(HLSLNoInterpolationModifier)
- };
+};
 
 
 // HLSL `noperspective` modifier
 class HLSLNoPerspectiveModifier : public InterpolationModeModifier 
 {
     SLANG_CLASS(HLSLNoPerspectiveModifier)
- };
+};
 
 
 // HLSL `linear` modifier
 class HLSLLinearModifier : public InterpolationModeModifier 
 {
     SLANG_CLASS(HLSLLinearModifier)
- };
+};
 
 
 // HLSL `sample` modifier
 class HLSLSampleModifier : public InterpolationModeModifier 
 {
     SLANG_CLASS(HLSLSampleModifier)
- };
+};
 
 
 // HLSL `centroid` modifier
 class HLSLCentroidModifier : public InterpolationModeModifier 
 {
     SLANG_CLASS(HLSLCentroidModifier)
- };
+};
 
 
 // HLSL `precise` modifier
 class PreciseModifier : public Modifier 
 {
     SLANG_CLASS(PreciseModifier)
- };
+};
 
 
 // HLSL `shared` modifier (which is used by the effect system,
@@ -474,14 +477,14 @@ class PreciseModifier : public Modifier
 class HLSLEffectSharedModifier : public Modifier 
 {
     SLANG_CLASS(HLSLEffectSharedModifier)
- };
+};
 
 
 // HLSL `groupshared` modifier
 class HLSLGroupSharedModifier : public Modifier 
 {
     SLANG_CLASS(HLSLGroupSharedModifier)
- };
+};
 
 
 // HLSL `static` modifier (probably doesn't need to be
@@ -489,7 +492,7 @@ class HLSLGroupSharedModifier : public Modifier
 class HLSLStaticModifier : public Modifier 
 {
     SLANG_CLASS(HLSLStaticModifier)
- };
+};
 
 
 // HLSL `uniform` modifier (distinct meaning from GLSL
@@ -497,14 +500,14 @@ class HLSLStaticModifier : public Modifier
 class HLSLUniformModifier : public Modifier 
 {
     SLANG_CLASS(HLSLUniformModifier)
- };
+};
 
 
 // HLSL `volatile` modifier (ignored)
 class HLSLVolatileModifier : public Modifier 
 {
     SLANG_CLASS(HLSLVolatileModifier)
- };
+};
 
 
 class AttributeTargetModifier : public Modifier 
@@ -513,7 +516,7 @@ class AttributeTargetModifier : public Modifier
  
     // A class to which the declared attribute type is applicable
     SyntaxClass<RefObject> syntaxClass;
- };
+};
 
 // Base class for checked and unchecked `[name(arg0, ...)]` style attribute.
 class AttributeBase : public Modifier 
@@ -521,7 +524,7 @@ class AttributeBase : public Modifier
     SLANG_CLASS(AttributeBase)
  
     List<RefPtr<Expr>> args;
- };
+};
 
 // A `[name(...)]` attribute that hasn't undergone any semantic analysis.
 // After analysis, this will be transformed into a more specific case.
@@ -530,7 +533,7 @@ class UncheckedAttribute : public AttributeBase
     SLANG_CLASS(UncheckedAttribute)
  
     RefPtr<Scope> scope;
- };
+};
 
 // A `[name(arg0, ...)]` style attribute that has been validated.
 class Attribute : public AttributeBase 
@@ -538,20 +541,20 @@ class Attribute : public AttributeBase
     SLANG_CLASS(Attribute)
  
     AttributeArgumentValueDict intArgVals;
- };
+};
 
 class UserDefinedAttribute : public Attribute 
 {
     SLANG_CLASS(UserDefinedAttribute)
  
- };
+};
 
 class AttributeUsageAttribute : public Attribute 
 {
     SLANG_CLASS(AttributeUsageAttribute)
  
     SyntaxClass<RefObject> targetSyntaxClass;
- };
+};
 
 // An `[unroll]` or `[unroll(count)]` attribute
 class UnrollAttribute : public Attribute 
@@ -559,42 +562,42 @@ class UnrollAttribute : public Attribute
     SLANG_CLASS(UnrollAttribute)
  
     IntegerLiteralValue getCount();
- };
+};
 
 class LoopAttribute : public Attribute 
 {
     SLANG_CLASS(LoopAttribute)
- };
+};
                // `[loop]`
 class FastOptAttribute : public Attribute 
 {
     SLANG_CLASS(FastOptAttribute)
- };
+};
             // `[fastopt]`
 class AllowUAVConditionAttribute : public Attribute 
 {
     SLANG_CLASS(AllowUAVConditionAttribute)
- };
+};
   // `[allow_uav_condition]`
 class BranchAttribute : public Attribute 
 {
     SLANG_CLASS(BranchAttribute)
- };
+};
              // `[branch]`
 class FlattenAttribute : public Attribute 
 {
     SLANG_CLASS(FlattenAttribute)
- };
+};
             // `[flatten]`
 class ForceCaseAttribute : public Attribute 
 {
     SLANG_CLASS(ForceCaseAttribute)
- };
+};
           // `[forcecase]`
 class CallAttribute : public Attribute 
 {
     SLANG_CLASS(CallAttribute)
- };
+};
                // `[call]`
 
 
@@ -602,14 +605,14 @@ class CallAttribute : public Attribute
 class PushConstantAttribute : public Attribute 
 {
     SLANG_CLASS(PushConstantAttribute)
- };
+};
 
 
 // [[vk_shader_record]] [[shader_record]]
 class ShaderRecordAttribute : public Attribute 
 {
     SLANG_CLASS(ShaderRecordAttribute)
- };
+};
 
 
 // [[vk_binding]]
@@ -619,27 +622,27 @@ class GLSLBindingAttribute : public Attribute
  
     int32_t binding = 0;
     int32_t set = 0;
- };
+};
 
 class GLSLSimpleIntegerLayoutAttribute : public Attribute 
 {
     SLANG_CLASS(GLSLSimpleIntegerLayoutAttribute)
  
     int32_t value = 0;
- };
+};
 
 // [[vk_location]]
 class GLSLLocationAttribute : public GLSLSimpleIntegerLayoutAttribute 
 {
     SLANG_CLASS(GLSLLocationAttribute)
- };
+};
 
 
 // [[vk_index]]
 class GLSLIndexAttribute : public GLSLSimpleIntegerLayoutAttribute 
 {
     SLANG_CLASS(GLSLIndexAttribute)
- };
+};
 
 
 // TODO: for attributes that take arguments, the syntax node
@@ -648,39 +651,39 @@ class GLSLIndexAttribute : public GLSLSimpleIntegerLayoutAttribute
 class MaxTessFactorAttribute : public Attribute 
 {
     SLANG_CLASS(MaxTessFactorAttribute)
- };
+};
 
 class OutputControlPointsAttribute : public Attribute 
 {
     SLANG_CLASS(OutputControlPointsAttribute)
- };
+};
 
 class OutputTopologyAttribute : public Attribute 
 {
     SLANG_CLASS(OutputTopologyAttribute)
- };
+};
 
 class PartitioningAttribute : public Attribute 
 {
     SLANG_CLASS(PartitioningAttribute)
- };
+};
 
 class PatchConstantFuncAttribute : public Attribute 
 {
     SLANG_CLASS(PatchConstantFuncAttribute)
  
     RefPtr<FuncDecl> patchConstantFuncDecl;
- };
+};
 class DomainAttribute : public Attribute 
 {
     SLANG_CLASS(DomainAttribute)
- };
+};
 
 
 class EarlyDepthStencilAttribute : public Attribute 
 {
     SLANG_CLASS(EarlyDepthStencilAttribute)
- };
+};
   // `[earlydepthstencil]`
 
 // An HLSL `[numthreads(x,y,z)]` attribute
@@ -695,7 +698,7 @@ class NumThreadsAttribute : public Attribute
     int32_t x;
     int32_t y;
     int32_t z;
- };
+};
 
 class MaxVertexCountAttribute : public Attribute 
 {
@@ -706,7 +709,7 @@ class MaxVertexCountAttribute : public Attribute
     // TODO: This should be an accessor that uses the
     // ordinary `args` list, rather than side data.
     int32_t value;
- };
+};
 
 class InstanceAttribute : public Attribute 
 {
@@ -717,7 +720,7 @@ class InstanceAttribute : public Attribute
     // TODO: This should be an accessor that uses the
     // ordinary `args` list, rather than side data.
     int32_t value;
- };
+};
 
 // A `[shader("stageName")]` attribute, which marks an entry point
 // to be compiled, and specifies the stage for that entry point
@@ -730,7 +733,7 @@ class EntryPointAttribute : public Attribute
     // TODO: This should be an accessor that uses the
     // ordinary `args` list, rather than side data.
     Stage stage;;
- };
+};
 
 // A `[__vulkanRayPayload]` attribute, which is used in the
 // standard library implementation to indicate that a variable
@@ -739,7 +742,7 @@ class EntryPointAttribute : public Attribute
 class VulkanRayPayloadAttribute : public Attribute 
 {
     SLANG_CLASS(VulkanRayPayloadAttribute)
- };
+};
 
 
 // A `[__vulkanCallablePayload]` attribute, which is used in the
@@ -749,7 +752,7 @@ class VulkanRayPayloadAttribute : public Attribute
 class VulkanCallablePayloadAttribute : public Attribute 
 {
     SLANG_CLASS(VulkanCallablePayloadAttribute)
- };
+};
 
 
 // A `[__vulkanHitAttributes]` attribute, which is used in the
@@ -759,7 +762,7 @@ class VulkanCallablePayloadAttribute : public Attribute
 class VulkanHitAttributesAttribute : public Attribute 
 {
     SLANG_CLASS(VulkanHitAttributesAttribute)
- };
+};
 
 
 // A `[mutating]` attribute, which indicates that a member
@@ -769,7 +772,7 @@ class VulkanHitAttributesAttribute : public Attribute
 class MutatingAttribute : public Attribute 
 {
     SLANG_CLASS(MutatingAttribute)
- };
+};
 
 
 // A `[__readNone]` attribute, which indicates that a function
@@ -780,7 +783,7 @@ class MutatingAttribute : public Attribute
 class ReadNoneAttribute : public Attribute 
 {
     SLANG_CLASS(ReadNoneAttribute)
- };
+};
 
 
 
@@ -788,32 +791,32 @@ class ReadNoneAttribute : public Attribute
 class HLSLGeometryShaderInputPrimitiveTypeModifier : public Modifier 
 {
     SLANG_CLASS(HLSLGeometryShaderInputPrimitiveTypeModifier)
- };
+};
 
 class HLSLPointModifier : public HLSLGeometryShaderInputPrimitiveTypeModifier 
 {
     SLANG_CLASS(HLSLPointModifier)
- };
+};
 
 class HLSLLineModifier : public HLSLGeometryShaderInputPrimitiveTypeModifier 
 {
     SLANG_CLASS(HLSLLineModifier)
- };
+};
 
 class HLSLTriangleModifier : public HLSLGeometryShaderInputPrimitiveTypeModifier 
 {
     SLANG_CLASS(HLSLTriangleModifier)
- };
+};
 
 class HLSLLineAdjModifier : public HLSLGeometryShaderInputPrimitiveTypeModifier 
 {
     SLANG_CLASS(HLSLLineAdjModifier)
- };
+};
 
 class HLSLTriangleAdjModifier : public HLSLGeometryShaderInputPrimitiveTypeModifier 
 {
     SLANG_CLASS(HLSLTriangleAdjModifier)
- };
+};
 
 
 // A modifier to be attached to syntax after we've computed layout
@@ -822,7 +825,7 @@ class ComputedLayoutModifier : public Modifier
     SLANG_CLASS(ComputedLayoutModifier)
  
     RefPtr<Layout> layout;
- };
+};
 
 
 class TupleVarModifier : public Modifier 
@@ -830,7 +833,7 @@ class TupleVarModifier : public Modifier
     SLANG_CLASS(TupleVarModifier)
  
 //  TupleFieldModifier* tupleField = nullptr;
- };
+};
 
 // A modifier to indicate that a constructor/initializer can be used
 // to perform implicit type conversion, and to specify the cost of
@@ -841,21 +844,21 @@ class ImplicitConversionModifier : public Modifier
  
     // The conversion cost, used to rank conversions
     ConversionCost cost;
- };
+};
 
 class FormatAttribute : public Attribute 
 {
     SLANG_CLASS(FormatAttribute)
  
     ImageFormat format;
- };
+};
 
 class AllowAttribute : public Attribute 
 {
     SLANG_CLASS(AllowAttribute)
  
     DiagnosticInfo const* diagnostic = nullptr;
- };
+};
 
 
 // A `[__extern]` attribute, which indicates that a function/type is defined externally
@@ -863,7 +866,7 @@ class AllowAttribute : public Attribute
 class ExternAttribute : public Attribute 
 {
     SLANG_CLASS(ExternAttribute)
- };
+};
 
 
 // An `[__unsafeForceInlineExternal]` attribute indicates that the callee should be inlined
@@ -872,6 +875,10 @@ class ExternAttribute : public Attribute
 class UnsafeForceInlineEarlyAttribute : public Attribute 
 {
     SLANG_CLASS(UnsafeForceInlineEarlyAttribute)
- };
+};
+
+
+#undef SLANG_ABSTRACT_CLASS
+#undef SLANG_CLASS
 
 } // namespace Slang
