+++ source/slang/slang-ast-val.h
@@ -6,6 +6,9 @@
 
 namespace Slang {
 
+#define SLANG_ABSTRACT_CLASS(x) SLANG_ABSTRACT_CLASS_REFLECT(x)
+#define SLANG_CLASS(x) SLANG_CLASS_REFLECT_WITH_ACCEPT(x)
+
 // Syntax class definitions for compile-time values.
 
 // A compile-time integer (may not have a specific concrete value)
@@ -200,5 +203,9 @@ class TaggedUnionSubtypeWitness : public SubtypeWitness
     virtual int GetHashCode() override;
     virtual RefPtr<Val> SubstituteImpl(SubstitutionSet subst, int * ioDiff) override;
 };
+
+
+#undef SLANG_ABSTRACT_CLASS
+#undef SLANG_CLASS
 
 } // namespace Slang
