+++ source/slang/slang-visitor.h
@@ -11,26 +11,49 @@
 
 namespace Slang {
 
+// Macros to generate from ast-generated-macro file the vistors
+
+// Only runs 'param' macro if the marker is NONE (ie not ABSTRACT here)
+#define SLANG_CLASS_ONLY_ABSTRACT(NAME, SUPER, ORIGIN, LAST, MARKER, TYPE, param)
+#define SLANG_CLASS_ONLY_NONE(NAME, SUPER, ORIGIN, LAST, MARKER, TYPE, param) param(NAME, SUPER, ORIGIN, LAST, MARKER, TYPE, param)
+
+#define SLANG_CLASS_ONLY(NAME, SUPER, ORIGIN, LAST, MARKER, TYPE, param) SLANG_CLASS_ONLY_##MARKER(NAME, SUPER, ORIGIN, LAST, MARKER, TYPE, param)
+
+// Dispatch decl
+#define SLANG_VISITOR_DISPATCH_DECL(NAME, SUPER, ORIGIN, LAST, MARKER, TYPE, param) \
+    virtual void dispatch_##NAME(NAME* obj, void* extra) = 0;
+
+// Dispatch
+
+#define SLANG_VISITOR_DISPATCH_RESULT_IMPL(NAME, SUPER, ORIGIN, LAST, MARKER, TYPE, param) \
+    virtual void dispatch_##NAME(NAME* obj, void* extra) override  { *(Result*)extra = ((Derived*) this)->visit##NAME(obj); }
+
+#define SLANG_VISITOR_DISPATCH_VOID_IMPL(NAME, SUPER, ORIGIN, LAST, MARKER, TYPE, param) \
+    virtual void dispatch_##NAME(NAME* obj, void*) override { ((Derived*) this)->visit##NAME(obj); }
+
+// Visitor with and without result
+
+#define SLANG_VISITOR_RESULT_VISIT_IMPL(NAME, SUPER, ORIGIN, LAST, MARKER, TYPE, param) \
+    Result visit##NAME(NAME* obj) { return ((Derived*) this)->visit##SUPER(obj); } \
+
+#define SLANG_VISITOR_VOID_VISIT_IMPL(NAME, SUPER, ORIGIN, LAST, MARKER, TYPE, param) \
+    void visit##NAME(NAME* obj) { ((Derived*) this)->visit##SUPER(obj); }
+
+// Args
+
+#define SLANG_VISITOR_DISPATCH_ARG_IMPL(NAME, SUPER, ORIGIN, LAST, MARKER, TYPE, param) \
+    virtual void dispatch_##NAME(NAME* obj, void* arg) override { ((Derived*) this)->visit##NAME(obj, *(Arg*)arg); }
+
+#define SLANG_VISITOR_VOID_VISIT_ARG_IMPL(NAME, SUPER, ORIGIN, LAST, MARKER, TYPE, param) \
+    void visit##NAME(NAME* obj, Arg const& arg) { ((Derived*) this)->visit##SUPER(obj, arg); }
 //
 // type Visitors
 //
 
 struct ITypeVisitor
 {
-    SLANG_DERIVED_ASTNode_Type(SLANG_VISITOR_DISPATCH, _)
+    SLANG_CHILDREN_ASTNode_Type(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_DECL)
 };
-
-#define SLANG_VISITOR_DISPATCH_IMPL(NAME, SUPER, STYLE, NODE_STYLE, PARAM) \
-    virtual void dispatch_##NAME(NAME* obj, void* extra) override { *(Result*)extra = ((Derived*) this)->visit##NAME(obj); }
-
-#define SLANG_VISIT_IMPL(NAME, SUPER, STYLE, NODE_STYLE, PARAM) \
-    Result visit##NAME(NAME* obj) { return ((Derived*) this)->visit##SUPER(obj); }
-
-#define SLANG_VISITOR_DISPATCH_WITH_ARG(NAME, SUPER, STYLE, NODE_STYLE, PARAM) \
-    virtual void dispatch_##NAME(NAME* obj, void* arg) override { ((Derived*) this)->visit##NAME(obj, *(Arg*)arg); }
-
-#define SLANG_VISIT_WITH_ARG(NAME, SUPER, STYLE, NODE_STYLE, PARAM) \
-    void visit##NAME(NAME* obj, Arg const& arg) { ((Derived*) this)->visit##SUPER(obj, arg); }
 
 template<typename Derived, typename Result = void, typename Base = ITypeVisitor>
 struct TypeVisitor : Base
@@ -49,8 +72,8 @@ struct TypeVisitor : Base
         return result;
     }
 
-    SLANG_DERIVED_ASTNode_Type(SLANG_VISITOR_DISPATCH_IMPL, _)
-    SLANG_DERIVED_ASTNode_Type(SLANG_VISIT_IMPL, _)
+    SLANG_CHILDREN_ASTNode_Type(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_RESULT_IMPL)
+    SLANG_CHILDREN_ASTNode_Type(SLANG_VISITOR_RESULT_VISIT_IMPL, _)
 };
 
 template<typename Derived, typename Base>
@@ -66,11 +89,9 @@ struct TypeVisitor<Derived,void,Base> : Base
         type->accept(this, 0);
     }
 
-    SLANG_DERIVED_ASTNode_Type(SLANG_VISITOR_DISPATCH_IMPL, _)
-    SLANG_DERIVED_ASTNode_Type(SLANG_VISIT_IMPL, _)
+    SLANG_CHILDREN_ASTNode_Type(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_VOID_IMPL)
+    SLANG_CHILDREN_ASTNode_Type(SLANG_VISITOR_VOID_VISIT_IMPL, _)
 };
-
-
 
 template<typename Derived, typename Arg, typename Base = ITypeVisitor>
 struct TypeVisitorWithArg : Base
@@ -80,8 +101,8 @@ struct TypeVisitorWithArg : Base
         type->accept(this, (void*)&arg);
     }
 
-    SLANG_DERIVED_ASTNode_Type(SLANG_VISITOR_DISPATCH_WITH_ARG, _)
-    SLANG_DERIVED_ASTNode_Type(SLANG_VISIT_WITH_ARG, _)
+    SLANG_CHILDREN_ASTNode_Type(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_ARG_IMPL)
+    SLANG_CHILDREN_ASTNode_Type(SLANG_VISITOR_VOID_VISIT_ARG_IMPL, _)
 };
 
 //
@@ -90,7 +111,7 @@ struct TypeVisitorWithArg : Base
 
 struct IExprVisitor
 {
-    SLANG_DERIVED_ASTNode_Expr(SLANG_VISITOR_DISPATCH, _)
+    SLANG_CHILDREN_ASTNode_Expr(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_DECL)
 };
 
 template<typename Derived, typename Result = void>
@@ -103,8 +124,8 @@ struct ExprVisitor : IExprVisitor
         return result;
     }
 
-    SLANG_DERIVED_ASTNode_Expr(SLANG_VISITOR_DISPATCH_IMPL, _)
-    SLANG_DERIVED_ASTNode_Expr(SLANG_VISIT_IMPL, _)
+    SLANG_CHILDREN_ASTNode_Expr(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_RESULT_IMPL)
+    SLANG_CHILDREN_ASTNode_Expr(SLANG_VISITOR_RESULT_VISIT_IMPL, _)
 };
 
 template<typename Derived>
@@ -115,8 +136,8 @@ struct ExprVisitor<Derived,void> : IExprVisitor
         expr->accept(this, 0);
     }
 
-    SLANG_DERIVED_ASTNode_Expr(SLANG_VISITOR_DISPATCH_IMPL, _)
-    SLANG_DERIVED_ASTNode_Expr(SLANG_VISIT_IMPL, _)
+    SLANG_CHILDREN_ASTNode_Expr(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_VOID_IMPL)
+    SLANG_CHILDREN_ASTNode_Expr(SLANG_VISITOR_VOID_VISIT_IMPL, _)
 };
 
 template<typename Derived, typename Arg>
@@ -127,8 +148,8 @@ struct ExprVisitorWithArg : IExprVisitor
         obj->accept(this, (void*)&arg);
     }
 
-    SLANG_DERIVED_ASTNode_Expr(SLANG_VISITOR_DISPATCH_WITH_ARG, _)
-    SLANG_DERIVED_ASTNode_Expr(SLANG_VISIT_WITH_ARG, _)
+    SLANG_CHILDREN_ASTNode_Expr(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_ARG_IMPL)
+    SLANG_CHILDREN_ASTNode_Expr(SLANG_VISITOR_VOID_VISIT_ARG_IMPL, _)
 };
 
 //
@@ -137,7 +158,7 @@ struct ExprVisitorWithArg : IExprVisitor
 
 struct IStmtVisitor
 {
-    SLANG_DERIVED_ASTNode_Stmt(SLANG_VISITOR_DISPATCH, _)
+    SLANG_CHILDREN_ASTNode_Stmt(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_DECL)
 };
 
 template<typename Derived, typename Result = void>
@@ -150,8 +171,8 @@ struct StmtVisitor : IStmtVisitor
         return result;
     }
 
-    SLANG_DERIVED_ASTNode_Stmt(SLANG_VISITOR_DISPATCH_IMPL, _)
-    SLANG_DERIVED_ASTNode_Stmt(SLANG_VISIT_IMPL, _)
+    SLANG_CHILDREN_ASTNode_Stmt(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_RESULT_IMPL)
+    SLANG_CHILDREN_ASTNode_Stmt(SLANG_VISITOR_RESULT_VISIT_IMPL, _)
 };
 
 template<typename Derived>
@@ -162,8 +183,8 @@ struct StmtVisitor<Derived,void> : IStmtVisitor
         stmt->accept(this, 0);
     }
 
-    SLANG_DERIVED_ASTNode_Stmt(SLANG_VISITOR_DISPATCH_IMPL, _)
-    SLANG_DERIVED_ASTNode_Stmt(SLANG_VISIT_IMPL, _)
+    SLANG_CHILDREN_ASTNode_Stmt(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_VOID_IMPL)
+    SLANG_CHILDREN_ASTNode_Stmt(SLANG_VISITOR_VOID_VISIT_IMPL, _)
 };
 
 //
@@ -172,7 +193,7 @@ struct StmtVisitor<Derived,void> : IStmtVisitor
 
 struct IDeclVisitor
 {
-    SLANG_DERIVED_ASTNode_Decl(SLANG_VISITOR_DISPATCH, _)
+    SLANG_CHILDREN_ASTNode_Decl(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_DECL)
 };
 
 template<typename Derived, typename Result = void>
@@ -185,8 +206,8 @@ struct DeclVisitor : IDeclVisitor
         return result;
     }
 
-    SLANG_DERIVED_ASTNode_Decl(SLANG_VISITOR_DISPATCH_IMPL, _)
-    SLANG_DERIVED_ASTNode_Decl(SLANG_VISIT_IMPL, _)
+    SLANG_CHILDREN_ASTNode_Decl(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_RESULT_IMPL)
+    SLANG_CHILDREN_ASTNode_Decl(SLANG_VISITOR_RESULT_VISIT_IMPL, _)
 };
 
 template<typename Derived>
@@ -197,8 +218,8 @@ struct DeclVisitor<Derived,void> : IDeclVisitor
         decl->accept(this, 0);
     }
 
-    SLANG_DERIVED_ASTNode_Decl(SLANG_VISITOR_DISPATCH_IMPL, _)
-    SLANG_DERIVED_ASTNode_Decl(SLANG_VISIT_IMPL, _)
+    SLANG_CHILDREN_ASTNode_Decl(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_VOID_IMPL)
+    SLANG_CHILDREN_ASTNode_Decl(SLANG_VISITOR_VOID_VISIT_IMPL, _)
 };
 
 template<typename Derived, typename Arg>
@@ -209,8 +230,8 @@ struct DeclVisitorWithArg : IDeclVisitor
         obj->accept(this, (void*)&arg);
     }
 
-    SLANG_DERIVED_ASTNode_Decl(SLANG_VISITOR_DISPATCH_WITH_ARG, _)
-    SLANG_DERIVED_ASTNode_Decl(SLANG_VISIT_WITH_ARG, _)
+    SLANG_CHILDREN_ASTNode_Decl(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_ARG_IMPL)
+    SLANG_CHILDREN_ASTNode_Decl(SLANG_VISITOR_VOID_VISIT_ARG_IMPL, _)
 };
 
 
@@ -220,7 +241,7 @@ struct DeclVisitorWithArg : IDeclVisitor
 
 struct IModifierVisitor
 {
-    SLANG_DERIVED_ASTNode_Modifier(SLANG_VISITOR_DISPATCH, _)
+    SLANG_CHILDREN_ASTNode_Modifier(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_DECL)
 };
 
 template<typename Derived, typename Result = void>
@@ -233,8 +254,8 @@ struct ModifierVisitor : IModifierVisitor
         return result;
     }
 
-    SLANG_DERIVED_ASTNode_Modifier(SLANG_VISITOR_DISPATCH_IMPL, _)
-    SLANG_DERIVED_ASTNode_Modifier(SLANG_VISIT_IMPL, _)
+    SLANG_CHILDREN_ASTNode_Modifier(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_RESULT_IMPL)
+    SLANG_CHILDREN_ASTNode_Modifier(SLANG_VISITOR_RESULT_VISIT_IMPL, _)
 };
 
 template<typename Derived>
@@ -245,8 +266,8 @@ struct ModifierVisitor<Derived, void> : IModifierVisitor
         modifier->accept(this, 0);
     }
 
-    SLANG_DERIVED_ASTNode_Modifier(SLANG_VISITOR_DISPATCH_IMPL, _)
-    SLANG_DERIVED_ASTNode_Modifier(SLANG_VISIT_IMPL, _)
+    SLANG_CHILDREN_ASTNode_Modifier(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_VOID_IMPL)
+    SLANG_CHILDREN_ASTNode_Modifier(SLANG_VISITOR_VOID_VISIT_IMPL, _)
 };
 
 //
@@ -255,7 +276,7 @@ struct ModifierVisitor<Derived, void> : IModifierVisitor
 
 struct IValVisitor : ITypeVisitor
 {
-    SLANG_DERIVED_ASTNode_Val(SLANG_VISITOR_DISPATCH, _)
+    SLANG_CHILDREN_ASTNode_Val(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_DECL)
 };
 
 template<typename Derived, typename Result = void, typename TypeResult = void>
@@ -268,8 +289,8 @@ struct ValVisitor : TypeVisitor<Derived, TypeResult, IValVisitor>
         return result;
     }
 
-    SLANG_DERIVED_ASTNode_Val(SLANG_VISITOR_DISPATCH_IMPL, _)
-    SLANG_DERIVED_ASTNode_Val(SLANG_VISIT_IMPL, _)
+    SLANG_CHILDREN_ASTNode_Val(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_RESULT_IMPL)
+    SLANG_CHILDREN_ASTNode_Val(SLANG_VISITOR_RESULT_VISIT_IMPL, _)
 };
 
 template<typename Derived>
@@ -280,8 +301,8 @@ struct ValVisitor<Derived, void, void> : TypeVisitor<Derived, void, IValVisitor>
         val->accept(this, 0);
     }
 
-    SLANG_DERIVED_ASTNode_Val(SLANG_VISITOR_DISPATCH_IMPL, _)
-    SLANG_DERIVED_ASTNode_Val(SLANG_VISIT_IMPL, _)
+    SLANG_CHILDREN_ASTNode_Val(SLANG_CLASS_ONLY, SLANG_VISITOR_DISPATCH_VOID_IMPL)
+    SLANG_CHILDREN_ASTNode_Val(SLANG_VISITOR_VOID_VISIT_IMPL, _)
 };
 
 }
